Move node 0.10 patching to linter.js
